services/telegram-bot/requirements.txt
python-telegram-bot==21.7
redis==5.0.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-dotenv==1.0.0
httpx==0.27.0
